[bgpd] Trivial fix of printf format/arg mismatch
2006-09-19 Andrew J. Schorr <ajschorr@alumni.princeton.edu>

	* bgpd.c: (peer_uptime) Fix printf format/arg mismatch in
	  zlog_warn message (%ld/size_t -> %lu/u_long).

+++ bgpd/bgpd.c
@@ -4250,8 +4250,7 @@ peer_uptime (time_t uptime2, char *buf, size_t len)
   /* Check buffer length. */
   if (len < BGP_UPTIME_LEN)
     {
-      /* XXX: warning: long int format, size_t arg (arg 2) */
-      zlog_warn ("peer_uptime (): buffer shortage %ld", len);
+      zlog_warn ("peer_uptime (): buffer shortage %lu", (u_long)len);
       /* XXX: should return status instead of buf... */
       snprintf (buf, len, "<error> "); 
       return buf;
